Add Google's sample app for adding Generic passes to Google Wallet
See:
- codelab - https://codelabs.developers.google.com/add-to-wallet-android
- workshop - https://www.youtube.com/watch?v=iZz_8N9WPVA
- oss project - https://github.com/google-wallet/android-codelab

To have Google Smart Tap working ensure you are having/creating a
Google Wallet Business Profile with Smart Tap enabled.
And then add the necessary details for all TODOs in the project.

For simpler functionality this app will not communicate with specific
issuer's wallet to download passes but will use any that is added as
JWT to `wallet_object_jwt.token` to add it to the Google Wallet.